requirements.txt
transformers
datasets
rank_bm25
sentence_transformers
pyvi
underthesea
sentencepiece
protobuf
gdown
beautifulsoup4
datasets
fastapi
numpy
ollama
pydantic
python-dotenv
pyvi
rank_bm25
Requests
selenium
sentence_transformers
underthesea
uvicorn
openai
chromadb
sse-starlette